clinical_trial_agent/requirements.txt
streamlit==1.32.0
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-groq==0.0.1
pandas==2.2.1
python-dotenv==1.0.1
pydantic==2.6.4
faiss-cpu
sentence-transformers
numpy
python-docx
